flaskapp/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
bitarray==2.9.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Cython==3.0.8
datasets==2.16.1
decorator==5.1.1
dill==0.3.7
Distance==0.1.3
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fairseq==0.12.2
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
frozenlist==1.4.1
fsspec==2023.10.0
g2p-en==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
hydra-core==1.0.7
idna==3.6
importlib-metadata==7.0.1
inflect==7.0.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lxml==5.1.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.3
omegaconf==2.0.6
openai==1.10.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
pooch==1.8.0
portalocker==2.8.2
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sacrebleu==2.4.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.15.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
